% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_normalization.R
\name{data_normalization}
\alias{data_normalization}
\title{transforms data into (0,1) interval}
\usage{
data_normalization(data)
}
\arguments{
\item{"data"}{A gene expression matrix with rows as samples and columns as genes.}
}
\value{
normalized gene expression data
}
\description{
This function transforms gene expression data into (0,1) interval
}
