% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wpLogicNet.R
\name{wpLogicNet}
\alias{wpLogicNet}
\title{structure and logic gate inference}
\usage{
wpLogicNet(
  data = expression_matrix,
  density = density,
  scale = scale,
  number_of_em_iterations = NA,
  BF_threshold = NA,
  max_num_regulators = NA
)
}
\arguments{
\item{"data"}{A gene expression matrix with rows as samples and columns as genes. Gene names must be given in the columns.}

\item{"density"}{Fitted density, "norm", "laplace", "logis", "cauchy"}

\item{"scale"}{Scale parameter of the fitted density}

\item{"number_of_em_iterations"}{number of em iterations}

\item{"BF_threshold"}{Bayes Factor threshold}

\item{"max_num_regulators"}{Max number of genes in a logic gate to regulate the target gene profile}
}
\value{
the predicted gene-gene interactions (directed edges) and logic gates.
"predicted_edges_top_logics" and "predicted_gates_top_logic" are prediction in top-logics mode.
"a_posteriori_distributed_sample" outputs the number of samples emitted from each active partition in top-logics mode (inferred a posteriori).
"top_edges" are predicted edges in top_edges mode, with column 1 as originating gene, column 2 as target gene and column 3 as the BF confidence score.
}
\description{
This function calls structure together with logic gates in gene regulatory networks with un-known structure. For this aim, it takes in steady-state gene expression profiles in (0,1) interval.
}
