% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wpLogicNet_structure_aware.R
\name{wpLogicNet_structure_aware}
\alias{wpLogicNet_structure_aware}
\title{logic gate inference in structure-aware mode}
\usage{
wpLogicNet_structure_aware(
  data = expression_matrix,
  density = density,
  scale = scale,
  true_edges = true_edges,
  number_of_em_iterations = NA
)
}
\arguments{
\item{"data"}{A gene expression matrix with rows as samples and columns as genes. Gene names must be given in columns.}

\item{"density"}{Fitted density, "norm", "laplace", "logis", "cauchy"}

\item{"scale"}{Scale parameter of the fitted density}

\item{"true_edges"}{A two-column matrix representing the true network topology. The first column for the originating and the second column for the target nodes.}

\item{"number_of_em_iterations"}{number of em iterations}
}
\value{
Outputs predicted logic gates.
}
\description{
This function calls logic gates in gene regulatory networks with a priori known structure (directed edges). For this aim, it takes in steady-state gene expression profiles in (0,1) interval.
}
